ipcache, daemon: Handle CIDR restoration from ipcache v1
The ipcache map was renamed to cilium_ipcache_v2 to distinguish it from
its previous version, which had a different value size. Therefore, on
upgrades from a previous minor version, the v2 ipcache map will not
exist.

This is problematic for CIDR policies because we rely on the ipcache map
being there during the upgrade to restore CIDR identities and avoid
packet drops. New logic is therefore needed to fallback to a restoration
from the v1 map if the v2 map isn't found.

We will thus first attempt to restore from the v2 ipcache (for agent
restarts and future upgrades) and, only if the v2 map isn't found, we
will fallback to an attempt to restore from the v1 ipcache (for the
current upgrade).

This new logic can be removed in the next minor version.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>