bpf:refactor: move host_secctx_from_ipcache to node config
From network-facing devices we use this config variable to decide whether
to lookup the secctx from ipcache. Given that we will transition to a
dedicated from-wireguard program in the subsequent commits, it will share
this codepath with bpf_host. So rather than duplicating the code, simply
move the secctx variable into NodeConfig: its values is decided in the
agent depending on the legacy routing value, but it is always set to
false for bpf_overlay, bpf_lxc, and bpf_host.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>